get api/v1/mongo/extremeWeather
Server is only started if there is a connection to the database. The ExtremeWeather events are queried (POST "api/v1/dwd/extremeWeather") and stored in the database (POST "api/v1/mongo/extremeWeather"). On the client side, a GET request is made to "api/v1/mongo/extremeWeather" to get the current weather data.